init(governance-adapter-nx): Create Nx adapter package boundary inside Plugins
