+++ packages/governance-adapter-nx/src/capability.spec.ts
