+++ packages/governance/package.json
@@ -78,6 +78,7 @@
     ]
   },
   "dependencies": {
+    "@anarchitects/governance-adapter-nx": "workspace:*",
     "minimatch": "9.0.5",
     "tslib": "^2.6.2",
     "typescript": "~5.9.2",
