+++ packages/governance/src/nx-adapter/capability.ts
@@ -1,39 +1,5 @@
-import type { GovernanceCapability } from '../extensions/capabilities.js';
-
-import type { AdapterWorkspaceSnapshot } from './types.js';
-
-export interface GovernanceNxCapabilityProject {
-  name: string;
-  root: string;
-  type?: string;
-  tags: string[];
-  targets: string[];
-}
-
-export interface GovernanceNxCapabilityData {
-  workspaceRoot: string;
-  projects: GovernanceNxCapabilityProject[];
-}
-
-export function createNxCapability(input: {
-  workspaceRoot: string;
-  snapshot: AdapterWorkspaceSnapshot;
-}): GovernanceCapability<GovernanceNxCapabilityData> {
-  const projects = input.snapshot.projects
-    .map((project) => ({
-      name: project.name,
-      root: project.root,
-      ...(project.type ? { type: project.type } : {}),
-      tags: [...(project.tags ?? [])],
-      targets: [...(project.targets ?? [])].sort((a, b) => a.localeCompare(b)),
-    }))
-    .sort((a, b) => a.name.localeCompare(b.name));
-
-  return {
-    id: 'capability:nx',
-    data: {
-      workspaceRoot: input.workspaceRoot,
-      projects,
-    },
-  };
-}
+export { createNxCapability } from '@anarchitects/governance-adapter-nx';
+export type {
+  GovernanceNxCapabilityData,
+  GovernanceNxCapabilityProject,
+} from '@anarchitects/governance-adapter-nx';
