+++ packages/governance/src/nx-adapter/types.ts
@@ -1,22 +1,5 @@
-export interface AdapterProject {
-  name: string;
-  root: string;
-  type: string;
-  tags?: string[];
-  targets?: string[];
-  metadata: Record<string, unknown>;
-}
-
-export interface AdapterDependency {
-  source: string;
-  target: string;
-  type: string;
-  sourceFile?: string;
-}
-
-export interface AdapterWorkspaceSnapshot {
-  root: string;
-  projects: AdapterProject[];
-  dependencies: AdapterDependency[];
-  codeownersByProject: Record<string, string[]>;
-}
+export type {
+  AdapterDependency,
+  AdapterProject,
+  AdapterWorkspaceSnapshot,
+} from '@anarchitects/governance-adapter-nx';
